# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ODBUS Consumer APIs",
        "description": "L5 Swagger OpenApi description for ODBUS Consumer APIs",
        "version": "0.1"
    },
    "paths": {
...
        "/api/RecentSearch": {
            "post": {
                "tags": [
                    "Save Recent Search of a user"
                ],
                "summary": "Save Recent Search of a user",
                "description": "Save Recent Search details of a user",
                "operationId": "d8c6272f89429024e9015d79e446e66c",
                "parameters": [
                    {
                        "name": "users_id",
                        "in": "query",
                        "description": "users id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    },
                    {
                        "name": "source",
                        "in": "query",
                        "description": "source name",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bhubaneswar"
                        }
                    },
                    {
                        "name": "destination",
                        "in": "query",
                        "description": "destination Name",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Balasore"
                        }
                    },
                    {
                        "name": "journey_date",
                        "in": "query",
                        "description": "journey date",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "07-12-2021"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Save search details of a user"
                    },
                    "206": {
                        "description": "Validation error"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "No record found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "502": {
                        "description": "Bad gateway"
                    },
                    "503": {
                        "description": "Service unavailable"
                    },
                    "504": {
                        "description": "Gateway timeout"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/RecentSearch/{userId}": {
            "get": {
                "tags": [
                    "Get Recent Search details of a User"
                ],
                "summary": "Get Recent Search details of a User",
                "operationId": "4425ac89ce4044dbe53aaccae51a7b09",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "Get Recent Search details of a User",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Recent Search of a User"
                    },
                    "206": {
                        "description": "validation error"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "No record found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "502": {
                        "description": "Bad gateway"
                    },
                    "503": {
                        "description": "Service unavailable"
                    },
                    "504": {
                        "description": "Gateway timeout"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "name": "Token based",
                "in": "header",
                "bearerFormat": "AUTH0",
                "scheme": "bearer"
            }
        }
    }
}